requirements.txt
pyone
pyyaml